cli: fix for folder-like files
In listing.py, if a file ends with "/", we treat it as a DirEntry
representing a file (this is a decision we inherit from blobfile,
which is fairly eager about creating such files). But we do also
want rmtree to remove such files, so we probably shouldn't remove
them from listtree. This is the easiest thing to do.