Update path.py: return annotation bool not int
Investigating https://github.com/python/mypy/pull/17487, I decided to check out an exemplar mypy_primer to see what all the fuss is about. Since it's so easy to fix this, and so obviously correct, I thought I'd do a PR about it while I'm here. The functions return bool and so should be annotated accordingly, not with int.